+++ packages/tools/utils/src/modules/array-[object-object].js
@@ -123,7 +123,7 @@ import * as _ from 'lamb';
 	],
 }
  * @since 0.6.0
- * @see {@link module:@svizzle/utils/object-[object-object].applyFnMap|applyFnMap}
+ * @see {@link module:@svizzle/utils/object-[any-object].applyFnMap|applyFnMap}
  * @see {@link module:@svizzle/utils/object-[object-object].makeMergeAppliedFnMap|makeMergeAppliedFnMap}
  * @see {@link module:@svizzle/utils/object-[object-object].transformValues|transformValues}
  * @see {@link module:@svizzle/utils/object-[object-object].transformPaths|transformPaths}