+++ packages/tools/utils/CHANGELOG.md
@@ -1,10 +1,11 @@
 ## `@svizzle/utils` v0.16.0 (next)
 
-- add `arraySumWith`, `pluckPath`, `updateKeys`
+- added `arraySumWith`, `pluckPath`, `updateKeys`
+- moved `applyFnMap` to `src/modules/object-[any-object]`
 
 ## `@svizzle/utils` v0.15.0
 
-- add `isPromise`
+- added `isPromise`
 - updated some dependencies
 
 ## `@svizzle/utils` v0.14.0